explain: create an `OptimizerTrace::into_rows` helper
