+++ src/adapter/src/explain/optimizer_trace.rs
@@ -20,7 +20,9 @@ use mz_repr::explain::tracing::{DelegateSubscriber, PlanTrace, TraceEntry};
 use mz_repr::explain::{
     Explain, ExplainConfig, ExplainError, ExplainFormat, ExprHumanizer, UsedIndexes,
 };
-use mz_sql::plan::{HirRelationExpr, HirScalarExpr};
+use mz_repr::{Datum, Row};
+use mz_sql::plan::{self, HirRelationExpr, HirScalarExpr};
+use mz_sql_parser::ast::{ExplainStage, NamedPlan};
 use mz_transform::dataflow::DataflowMetainfo;
 use mz_transform::notice::RawOptimizerNotice;
 use tracing::dispatcher;
@@ -28,6 +30,7 @@ use tracing_subscriber::prelude::*;
 
 use crate::coord::peek::FastPathPlan;
 use crate::explain::Explainable;
+use crate::AdapterError;
 
 /// Provides functionality for tracing plans generated by the execution of an
 /// optimization pipeline.
@@ -89,6 +92,88 @@ impl OptimizerTrace {
 
             OptimizerTrace(dispatcher::Dispatch::new(subscriber))
         }
+    }
+
+    /// Convert the optimizer trace into a vector or rows that can be returned
+    /// to the client.
+    pub fn into_rows(
+        self,
+        format: ExplainFormat,
+        config: &ExplainConfig,
+        humanizer: &dyn ExprHumanizer,
+        row_set_finishing: Option<RowSetFinishing>,
+        used_indexes: UsedIndexes,
+        dataflow_metainfo: DataflowMetainfo,
+        stage: ExplainStage,
+        stmt_kind: plan::ExplaineeStatementKind,
+    ) -> Result<Vec<Row>, AdapterError> {
+        let mut trace = self.drain_all(
+            format,
+            config,
+            humanizer,
+            row_set_finishing,
+            used_indexes,
+            dataflow_metainfo,
+        )?;
+
+        let rows = match stage.path() {
+            None => {
+                // For the `Trace` (pseudo-)stage, return the entire trace as
+                // triples of (time, path, plan) values.
+                let rows = trace
+                    .into_iter()
+                    .map(|entry| {
+                        // The trace would have to take over 584 years to overflow a u64.
+                        let span_duration = u64::try_from(entry.span_duration.as_nanos());
+                        Row::pack_slice(&[
+                            Datum::from(span_duration.unwrap_or(u64::MAX)),
+                            Datum::from(entry.path.as_str()),
+                            Datum::from(entry.plan.as_str()),
+                        ])
+                    })
+                    .collect();
+                rows
+            }
+            Some(path) => {
+                // For everything else, return the plan for the stage identified
+                // by the corresponding path.
+
+                // A helper struct hat removes the first (and by assumption the
+                // only) plan that matches the given path from the collected
+                // trace.
+                let mut remove_plan = |path: &'static str| {
+                    let index = trace.iter().position(|entry| entry.path == path);
+                    index.map(|index| trace.remove(index))
+                };
+
+                // For certain stages we want to return the resulting fast path
+                // plan instead of the selected stage if it is present.
+                let plan = if stage.show_fast_path() && !config.no_fast_path {
+                    remove_plan(NamedPlan::FastPath.path()).or_else(|| remove_plan(path))
+                } else {
+                    remove_plan(path)
+                };
+
+                let row = plan
+                    .map(|entry| Row::pack_slice(&[Datum::from(entry.plan.as_str())]))
+                    .ok_or_else(|| {
+                        if !stmt_kind.supports(&stage) {
+                            // Print a nicer error for unsupported stages.
+                            AdapterError::Unstructured(anyhow::anyhow!(format!(
+                                "cannot EXPLAIN {stage} FOR {stmt_kind}"
+                            )))
+                        } else {
+                            // We don't expect this stage to be missing.
+                            AdapterError::Internal(format!(
+                                "stage `{path}` not present in the collected optimizer trace",
+                            ))
+                        }
+                    })?;
+                vec![row]
+            }
+        };
+
+        Ok(rows)
     }
 
     /// Collect all traced plans for all plan types `T` that are available in
